requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
python-dateutil==2.8.2
Werkzeug==2.0.1
PyJWT==2.3.0
bcrypt==3.2.0 